Change atoms to tsx funcitonal components
